backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
brotli==1.2.0
bs4==0.0.2
cachetools==6.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.9.0
distro==1.9.0
Django==4.2.11
dj-database-url==2.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
fonttools==4.61.1
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
joblib==1.5.3
kombu==5.6.1
numpy==2.4.0
oauthlib==3.3.1
openai==2.21.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.58.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pyphen==0.17.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==7.1.0
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.5
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.7.0
xgboost==3.1.2
zopfli==0.4.1
